076Streamlit/requirements.txt
numpy~=1.21
pandas~=1.0.5
streamlit~=1.4.0
matplotlib~=3.2.2
plotly~=4.14.3
Pillow~=9.0.0
streamlit-option-menu~=0.2.10